+++ js/shared/rk-version.js
@@ -1,1 +1,1 @@
-window.RK_VERSION = "v0.6.0";
+window.RK_VERSION = "v0.7.0";